requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
autopep8==1.5.4
build==1.2.2.post1
certifi==2020.6.20
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.0
click==8.1.8
cloudinary==1.41.0
colorama==0.4.6
cryptography==44.0.2
cssselect2==0.8.0
Cython==3.0.11
defusedxml==0.7.1
diff-match-patch==20241021
dj-database-url==0.5.0
Django==5.1.3
django-allauth==65.3.0
django-ckeditor==6.7.1
django-ckeditor-5==0.1.6
django-cloudinary-storage==0.2.3
django-crispy-forms==1.12.0
django-debug-toolbar==2.2
django-dotenv==1.4.2
django-environ==0.9.0
django-import-export==4.3.3
django-jazzmin==3.0.1
django-js-asset==2.2.0
django-social-share==2.2.1
django-static-fontawesome==5.14.0.0
django-taggit==3.0.0
django-tailwind==3.8.0
djangorestframework==3.15.2
gunicorn==20.0.4
html5lib==1.1
idna==2.10
lxml==5.3.1
mock==4.0.2
mysqlclient==2.2.6
oscrypto==1.3.0
packaging==24.2
pillow==11.0.0
pycodestyle==2.12.1
pycparser==2.22
pyHanko==0.26.0
pyhanko-certvalidator==0.26.7
PyMySQL==1.1.1
pypdf==5.3.1
pyproject_hooks==1.2.0
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-decouple==3.3
pytz==2020.1
PyYAML==6.0.2
qrcode==8.0
reportlab==4.3.1
requests==2.32.3
setuptools==75.6.0
six==1.16.0
sqlparse==0.3.1
stripe==11.6.0
svglib==1.5.1
tablib==3.7.0
tinycss2==1.4.0
toml==0.10.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.3.1
uritools==4.0.3
urllib3==2.2.3
webencodings==0.5.1
wheel==0.45.1
whitenoise==5.2.0
xhtml2pdf==0.2.17
